+++ src/lib/cherrypickAndCreateTargetPullRequest.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk';
 import { isEmpty, difference } from 'lodash';
-import { BackportError } from '../errors/BackportError';
 import { ValidConfigOptions } from '../options/options';
+import { BackportError } from './BackportError';
 import { spawnPromise } from './child-process-promisified';
 import { getRepoPath } from './env';
 import { getCommitsWithoutBackports } from './getCommitsWithoutBackports';
