+++ src/lib/github/v4/fetchCommits/fetchCommitBySha.ts
@@ -1,6 +1,6 @@
 import gql from 'graphql-tag';
-import { BackportError } from '../../../../errors/BackportError';
 import { ValidConfigOptions } from '../../../../options/options';
+import { BackportError } from '../../../BackportError';
 import { swallowMissingConfigFileException } from '../../../remoteConfig';
 import {
   Commit,
